Move DecimalArray to array_decimal.rs
