+++ arrow/src/array/mod.rs
@@ -146,6 +146,7 @@
 mod array;
 mod array_binary;
 mod array_boolean;
+mod array_decimal;
 mod array_dictionary;
 mod array_list;
 mod array_map;
@@ -177,10 +178,10 @@ pub use self::data::ArrayDataBuilder;
 pub use self::data::ArrayDataRef;
 
 pub use self::array_binary::BinaryArray;
-pub use self::array_binary::DecimalArray;
 pub use self::array_binary::FixedSizeBinaryArray;
 pub use self::array_binary::LargeBinaryArray;
 pub use self::array_boolean::BooleanArray;
+pub use self::array_decimal::DecimalArray;
 pub use self::array_dictionary::DictionaryArray;
 pub use self::array_list::FixedSizeListArray;
 pub use self::array_list::LargeListArray;
